+++ edit_tags.php
@@ -128,7 +128,7 @@ if (isset($_REQUEST['new'])) {
     ?>
     <h2>New Tag</h2>
     <script type="text/javascript" charset="utf-8">
-        $(document).ready(ask_before_exiting);
+        $(document).ready(lwt_form_check.askBeforeExit);
     </script> 
     <form name="newtag" class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>" method="post">
     <table class="tab1" cellspacing="0" cellpadding="5">
@@ -149,7 +149,7 @@ if (isset($_REQUEST['new'])) {
         </tr>
         <tr>
             <td class="td1 right" colspan="2">
-                <input type="button" value="Cancel" onclick="{resetDirty(); location.href='edit_tags.php';}" /> 
+                <input type="button" value="Cancel" onclick="{lwt_form_check.resetDirty(); location.href='edit_tags.php';}" /> 
                 <input type="submit" name="op" value="Save" />
             </td>
         </tr>
@@ -166,7 +166,7 @@ if (isset($_REQUEST['new'])) {
         ?>
      <h2>Edit Tag</h2>
      <script type="text/javascript" charset="utf-8">
-         $(document).ready(ask_before_exiting);
+         $(document).ready(lwt_form_check.askBeforeExit);
      </script>  
      <form name="edittag" class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>#rec<?php echo $_REQUEST['chg']; ?>" method="post">
      <input type="hidden" name="TgID" value="<?php echo $record['TgID']; ?>" />
@@ -187,7 +187,7 @@ if (isset($_REQUEST['new'])) {
         </tr>
         <tr>
         <td class="td1 right" colspan="2">
-            <input type="button" value="Cancel" onclick="{resetDirty(); location.href='edit_tags.php#rec<?php echo $_REQUEST['chg']; ?>';}" /> 
+            <input type="button" value="Cancel" onclick="{lwt_form_check.resetDirty(); location.href='edit_tags.php#rec<?php echo $_REQUEST['chg']; ?>';}" /> 
             <input type="submit" name="op" value="Change" /></td>
         </tr>
     </table>