+++ edit_word.php
@@ -285,7 +285,7 @@ function edit_word_do_form($wid, $text_id, $ord, $fromAnn)
     pagestart_nobody($titletext);
     ?>
 <script type="text/javascript">
-    $(document).ready(ask_before_exiting);
+    $(document).ready(lwt_form_check.askBeforeExit);
     $(window).on('beforeunload',function() {
         setTimeout(function() {window.parent.frames['ru'].location.href = 'empty.html';}, 0);
     });