+++ edit_words.php
@@ -678,7 +678,7 @@ if (isset($_REQUEST['new']) && isset($_REQUEST['lang'])) {
 
     <h2>New Term</h2>
     <script type="text/javascript" charset="utf-8">
-        $(document).ready(ask_before_exiting);
+        $(document).ready(lwt_form_check.askBeforeExit);
     </script>  
     <form name="newword" class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>" method="post">
     <input type="hidden" name="WoLgID" id="langfield" value="<?php echo $_REQUEST['lang']; ?>" />
@@ -726,7 +726,7 @@ if (isset($_REQUEST['new']) && isset($_REQUEST['lang'])) {
     'document.forms[\'newword\'].WoSentence', 
     'document.forms[\'newword\'].WoText'); ?>
         &nbsp; &nbsp;
-    <input type="button" value="Cancel" onclick="{resetDirty(); location.href='edit_words.php';}" /> 
+    <input type="button" value="Cancel" onclick="{lwt_form_check.resetDirty(); location.href='edit_words.php';}" /> 
     <input type="submit" name="op" value="Save" /></td>
     </tr>
     </table>
@@ -752,7 +752,7 @@ if (isset($_REQUEST['new']) && isset($_REQUEST['lang'])) {
     
      <h2>Edit Term</h2>
      <script type="text/javascript" charset="utf-8">
-         $(document).ready(ask_before_exiting);
+         $(document).ready(lwt_form_check.askBeforeExit);
      </script>
      <form name="editword" class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>#rec<?php echo $_REQUEST['chg']; ?>" method="post">
      <input type="hidden" name="WoID" value="<?php echo $record['WoID']; ?>" />
@@ -802,7 +802,7 @@ if (isset($_REQUEST['new']) && isset($_REQUEST['lang'])) {
         <td class="td1 right" colspan="2">  &nbsp;
             <?php echo createDictLinksInEditWin2($record['WoLgID'], 'document.forms[\'editword\'].WoSentence', 'document.forms[\'editword\'].WoText'); ?>
             &nbsp; &nbsp;
-            <input type="button" value="Cancel" onclick="{resetDirty(); location.href='edit_words.php#rec<?php echo $_REQUEST['chg']; ?>';}" /> 
+            <input type="button" value="Cancel" onclick="{lwt_form_check.resetDirty(); location.href='edit_words.php#rec<?php echo $_REQUEST['chg']; ?>';}" /> 
             <input type="submit" name="op" value="Change" />
         </td>
      </tr>