+++ settings.php
@@ -204,7 +204,7 @@ echo error_message_with_hide($message, true);
 
 ?>
 <script type="text/javascript" charset="utf-8">
-    $(document).ready(ask_before_exiting);
+    $(document).ready(lwt_form_check.askBeforeExit);
 </script>
 <form class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>" method="post">
 <table class="tab1" cellspacing="0" cellpadding="5">
@@ -728,10 +728,10 @@ echo error_message_with_hide($message, true);
     <tr>
         <td class="td1 right" colspan="4">
             <input type="button" value="&lt;&lt; Back" 
-            onclick="{resetDirty(); location.href='index.php';}" />
+            onclick="{lwt_form_check.resetDirty(); location.href='index.php';}" />
             <span class="nowrap"></span>
             <input type="button" value="Reset all settings to default" 
-            onclick="{resetDirty(); location.href='settings.php?op=reset';}" />
+            onclick="{lwt_form_check.resetDirty(); location.href='settings.php?op=reset';}" />
             <span class="nowrap"></span>
             <input type="submit" name="op" value="Save" />
         </td>